+++ objectbox/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * New `@Entity()` annotation field `type realClass` to support some custom code generators.
 * Fix `ToMany` showing duplicate items after adding them before reading the previous list.
+* Fix invalid native free during store shutdown if large data was inserted (more than 64 kilobytes flatbuffer).
 
 ## 1.0.0 (2021-05-18)