+++ objectbox/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## latest
 
 * New `@Entity()` annotation field `type realClass` to support some custom code generators.
+* Rename semi-internal `QueryRelationProperty` to `QueryRelationToOne` and `QueryRelationMany` to `QueryRelationToMany`
+  to help users choosing the right link function: `link()`/`linkMany()`.
 * Fix `ToMany` showing duplicate items after adding them before reading the previous list.
 * Fix invalid native free during store shutdown if large data was inserted (more than 64 kilobytes flatbuffer).